+++ src/main/resources/application.properties
@@ -23,7 +23,7 @@ spring.main.allow-circular-references=true
 spring.security.oauth2.client.registration.google.client-id=${GOOGLE_ID}
 spring.security.oauth2.client.registration.google.client-secret=${GOOGLE_SECRECT}
 
-
+spring.main.banner-mode=off
 spring.security.oauth2.client.registration.facebook.client-id=${FACEBOOK_ID} 
 spring.security.oauth2.client.registration.facebook.client-secret=${FACEBOOK_SECRECT}
